Move files to one dir
